requirements.txt
asgiref==3.8.1
chardet==5.2.0
Django==5.1.2
django-cors-headers==4.6.0
djangorestframework==3.15.2
pillow==11.0.0
psycopg2==2.9.10
python-dotenv==1.0.1
reportlab==4.2.5
sqlparse==0.5.1
typing_extensions==4.12.2
